// src/types/searchCity.ts
// types/weatherCity.ts
export type CityOptions = {
  name_startsWith: string;
  cities?: string;
};

export type GeonameCity = {
  adminCode1: string;
  lng: string;
  geonameId: number;
  toponymName: string;
  countryId: string;
  fcl: string;
  population: number;
  countryCode: string;
  name: string;
  fclName: string;
  adminCodes1?: {
    ISO3166_2: string;
  };
  countryName: string;
  fcodeName: string;
  adminName1: string;
  lat: string;
  fcode: string;
};

export type GeonamesResponse = {
  totalResultsCount: number;
  geonames: GeonameCity[];
};